+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -311,6 +311,7 @@ sub build_isf_public_interface_contract {
             't/1219-isf-rule-transaction-priority.t',
             't/1220-isf-arbitration-schedule-report.t',
             't/1221-isf-rule-expression-assignment.t',
+            't/1222-isf-rule-expression-conflict-report.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',